requirements.txt
streamlit>=1.28.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
h5py>=3.7.0
scikit-learn>=1.0.0
einops>=0.6.0
Pillow>=8.0.0 